# expression: make `baseBuiltinFunc` support converting between row-based evaluation and vectorized evaluation