<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
We will implement the vectorized evaluation for all our builtin functions soon.
For each builtin function, we only keep one kind of implementation (row-based or vectorized) since it's hard and unnecessary to maintain these two.

Therefore, we need a converter to make all builtin function support row-based evaluation and vectorized evaluation both since these two kinds of evaluation are needed by users.

We decide to make `baseBuiltinFunc` support this feature so that all builtin functions which inherit it can support this converting too.

### What is changed and how it works?
1. Make `baseBuiltinFunc` support this converting;
2. Add some unit tests;

Here is the benchmark: 
```
BenchmarkMockDoubleIntVec-12             3000000               427 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleIntRow-12              100000             14554 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleIntVec2Row-12           50000             32872 ns/op               0 B/op          0 allocs/op
BenchmarkMockDoubleIntRow2Vec-12          100000             12993 ns/op            8320 B/op          2 allocs/op
BenchmarkMockDoubleStrVec-12               50000             33466 ns/op              64 B/op          1 allocs/op
BenchmarkMockDoubleStrRow-12                2000            909000 ns/op          426348 B/op       4097 allocs/op
BenchmarkMockDoubleStrVec2Row-12           10000            105747 ns/op           65541 B/op       1024 allocs/op
BenchmarkMockDoubleStrRow2Vec-12          100000             13016 ns/op            8320 B/op          2 allocs/op
```
1. Vectorized evaluation is 30X than row-based evaluation;
2. Converting from Vec to Row has some performance drawback for fixed-length type like `Int`, but I think it's ok since we will finally change all places which is import for performance to vectorized evaluation;

**All changes are for this new feature and unit tests, so it's easy to review.**

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test